+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 4.8
 - Accessibility: add alt attribute to images in Bottom.php, General_Info.inc.php
 - Accessibility: use onfocus instead of onclick in Inputs.php
 - Accessibility: hidden input title using .a11y-hidden class in Inputs.php, Calendar.php, CalendarDay.inc.php
+- Accessibility: add hidden input label using .a11y-hidden class in Date.php
 
 Changes in 4.7
 --------------